update quote to use better randomness

+++ Backend/Models/Quote.cs
@@ -1,3 +1,3 @@
 ﻿namespace PDPWebsite.Models;
 
-public record Quote(Guid? Id, string Text, string Title, float? Chance, ulong Creator, ulong Target, DateTime CreatedAt, uint Color);
+public record Quote(Guid? Id, string Text, string Title, uint? Chance, ulong Creator, ulong Target, DateTime CreatedAt, uint Color);
